Rename suggester package names to Dev-Assist
